chore: move file for network stats method to lib/commands/network directory (from lib/commands/wallet)
