--- lib/commands/network/disconnect.js
@@ -1,7 +1,0 @@
-'use strict'
-
-const chalk = require('chalk')
-
-module.exports = () => {
-  console.log(chalk.blue('Hello World!'))
-}
