+++ lib/commands/network/stats.js
